Refactored prefilters
